move internal only project things to a new internal module
